+++ src/nvim/cursor_shape.h
@@ -44,9 +44,9 @@ typedef struct cursor_entry {
   CursorShape shape;      ///< cursor shape: one of the SHAPE_ defines
   int mshape;             ///< mouse shape: one of the MSHAPE defines
   int percentage;         ///< percentage of cell for bar
-  long blinkwait;         ///< blinking, wait time before blinking starts
-  long blinkon;           ///< blinking, on time
-  long blinkoff;          ///< blinking, off time
+  int blinkwait;          ///< blinking, wait time before blinking starts
+  int blinkon;            ///< blinking, on time
+  int blinkoff;           ///< blinking, off time
   int id;                 ///< highlight group ID
   int id_lm;              ///< highlight group ID for :lmap mode
   char *name;             ///< mode short name
